project/requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.3
billiard==4.1.0
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
Django==4.1.7
django-braces==1.15.0
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-environ==0.10.0
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.3.0
drf-yasg==1.21.5
flower==2.0.0
humanize==4.7.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mongoengine==0.27.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pymongo==4.4.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.2
redis==3.5.3
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.4.2
social-auth-storage-mongoengine==1.0.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
